ready to initiate

+++ app.py
@@ -4,7 +4,7 @@ from flask import Flask
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql://user:pass@postgresql/mydatabase'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql://admin:admin@postgresql/projectdb'
 db = SQLAlchemy(app)
 db.create_all()
 
@@ -22,13 +22,11 @@ def index():
     return response
 
 
-
-
-# reinicie a aplicação apos acessar esta rota
-@app.route('/dropall')
-def dropall():
-    from db import db
-    db.drop_all()
-
+def recreate_db(recreate=True):
+    if recreate:
+        from db import db
+        db.drop_all()
+        db.create_all()
 if __name__ == '__main__':
+    recreate_db(False)
     app.run(host='0.0.0.0',debug=True)
